Profit rate corrected
Fix profit/sec tracking after large money changes

Reworked profit rate calculation so it no longer depends on wallet-to-wallet delta comparisons. ScoreManager now tracks actual income awarded during each rate window, and ProfitRate consumes that tracked income to calculate the displayed profit per second.

This fixes a long-running issue where rapid large purchases or large money gains could cause the profit/sec display to overcorrect to $0/s and fail to recover, even while employee income was still being generated correctly.

Also updated profit storage from float to double for better precision at high money values, while keeping existing float getters for compatibility with older scripts. Save data and save/load handling were updated so large profit values keep their precision across sessions.

Profit/sec now counts real income sources including manual dish rewards, employee income, bubble cash rewards, bubble-related earning boosts, and background wash rewards. Purchases still subtract from total profit without counting as negative income.

+++ Assets/SaveData.cs
@@ -9,7 +9,7 @@ public class SaveData
     // Legacy single-kitchen fields. These are still mirrored from kitchen_1 so old debugging
     // tools or old fallback paths do not immediately break.
     public long totalDishes = 0;
-    public float totalProfit = 0f;
+    public double totalProfit = 0d;
     public int dishCountIncrement = 1;
     public float profitPerDish = 1f;
     public float dishProfitMultiplier = 1f;
@@ -47,7 +47,7 @@ public class KitchenSaveData
     public string kitchenId = "kitchen_1";
 
     public long totalDishes = 0;
-    public float totalProfit = 0f;
+    public double totalProfit = 0d;
     public int dishCountIncrement = 1;
     public float profitPerDish = 1f;
     public float dishProfitMultiplier = 1f;
